Changed the navigation code structure.
The navigation now consists of a set of coords (nav_xy (i32,i32)) and a
boolean locking system where a variable is set depending on which state
the UI should render in.
ui_data.deeper() when a new set of coords is needed (submenu entered)
and ui_data.surface() when a submenu is exited.

In spite of this the code to render the UI currently uses hard-coded
indexes of which ui_data.nav_xy to use, this due to entering/exiting
some sub-menu's from anywhere is possible.