Fixed the character view in a encounter.
Now a user can select a character and view a total of all attacks and
displaying the highest hitting of that attack. If the user enters
something in the filters however, all attacks matching the filter is
shown.
This does not include the scrolling code needed when displaying all
attacks or when the number of characters/attack-types exceed the window-
size.